Move isProxiable to guards file
