+++ packages/arbor-store/src/NodeHandler.ts
@@ -3,8 +3,7 @@ import NodeCache from "./NodeCache"
 import Path from "./Path"
 import Subscribers from "./Subscribers"
 import { NotAnArborNodeError } from "./errors"
-import { isNode } from "./guards"
-import isProxiable from "./isProxiable"
+import { isNode, isProxiable } from "./guards"
 
 const PROXY_HANDLER_API = ["apply", "get", "set", "deleteProperty"]
