+++ packages/arbor-store/src/guards.ts
@@ -10,3 +10,16 @@ export function isArborNode<T extends object>(
 ): value is ArborNode<T> {
   return isNode(value)
 }
+
+export const ArborProxiable = Symbol.for("ArborProxiable")
+
+export function isProxiable(value: unknown): value is object {
+  if (value == null) return false
+
+  return (
+    Array.isArray(value) ||
+    value instanceof Map ||
+    value[ArborProxiable] === true ||
+    value.constructor === Object
+  )
+}
