+++ packages/arbor-store/src/index.ts
@@ -16,7 +16,6 @@ export type { Record } from "./Repository"
 export { default as Subscribers } from "./Subscribers"
 export type { MutationEvent, Subscriber, Unsubscribe } from "./Subscribers"
 export * from "./errors"
-export { isArborNode, isNode } from "./guards"
-export { ArborProxiable, default as isProxiable } from "./isProxiable"
+export * from "./guards"
 export type { Mutation } from "./mutate"
 export * from "./utilities"
