--- packages/arbor-store/src/isProxiable.ts
@@ -1,12 +1,0 @@
-export const ArborProxiable = Symbol.for("ArborProxiable")
-
-export default function isProxiable(value: unknown): value is object {
-  if (value == null) return false
-
-  return (
-    Array.isArray(value) ||
-    value instanceof Map ||
-    value[ArborProxiable] === true ||
-    value.constructor === Object
-  )
-}
